LOG-1730: Refactoring CLO for new collector
