+++ internal/k8shandler/fluentd.go
@@ -3,27 +3,15 @@ package k8shandler
 import (
 	"context"
 	"fmt"
-	"io/ioutil"
-	"path"
-	"reflect"
 	"strconv"
-	"time"
-
-	"github.com/openshift/cluster-logging-operator/internal/utils/comparators/servicemonitor"
 
 	"github.com/ViaQ/logerr/log"
-	monitoringv1 "github.com/coreos/prometheus-operator/pkg/apis/monitoring/v1"
 	logging "github.com/openshift/cluster-logging-operator/apis/logging/v1"
-	"github.com/openshift/cluster-logging-operator/internal/components/fluentd"
 	"github.com/openshift/cluster-logging-operator/internal/constants"
-	"github.com/openshift/cluster-logging-operator/internal/factory"
 	"github.com/openshift/cluster-logging-operator/internal/utils"
 	"github.com/openshift/cluster-logging-operator/internal/utils/comparators/daemonsets"
-	"github.com/openshift/cluster-logging-operator/internal/utils/comparators/services"
 	"k8s.io/apimachinery/pkg/api/errors"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/types"
-	"k8s.io/apimachinery/pkg/util/intstr"
 	"k8s.io/apimachinery/pkg/util/sets"
 	"k8s.io/client-go/util/retry"
 
@@ -76,213 +64,6 @@ const (
 	tmpValue                   = "/tmp"
 )
 
-func (clusterRequest *ClusterLoggingRequest) removeCollector(name string) (err error) {
-	if clusterRequest.isManaged() {
-
-		if err = clusterRequest.RemoveService(name); err != nil {
-			return
-		}
-
-		if err = clusterRequest.RemoveServiceMonitor(name); err != nil {
-			return
-		}
-
-		if err = clusterRequest.RemovePrometheusRule(name); err != nil {
-			return
-		}
-
-		if err = clusterRequest.RemoveConfigMap(name); err != nil {
-			return
-		}
-
-		caName := fmt.Sprintf("%s-trusted-ca-bundle", name)
-		if err = clusterRequest.RemoveConfigMap(caName); err != nil {
-			return
-		}
-
-		if err = clusterRequest.RemoveDaemonset(name); err != nil {
-			return
-		}
-
-		// Wait longer than the terminationGracePeriodSeconds
-		time.Sleep(12 * time.Second)
-
-		if err = clusterRequest.RemoveSecret(name); err != nil {
-			return
-		}
-
-	}
-
-	return nil
-}
-
-func (clusterRequest *ClusterLoggingRequest) reconcileFluentdService() error {
-	desired := factory.NewService(
-		constants.CollectorName,
-		clusterRequest.Cluster.Namespace,
-		constants.CollectorName,
-		[]v1.ServicePort{
-			{
-				Port:       metricsPort,
-				TargetPort: intstr.FromString(metricsPortName),
-				Name:       metricsPortName,
-			},
-			{
-				Port:       exporterPort,
-				TargetPort: intstr.FromString(exporterPortName),
-				Name:       exporterPortName,
-			},
-		},
-	)
-
-	desired.Annotations = map[string]string{
-		"service.alpha.openshift.io/serving-cert-secret-name": constants.CollectorMetricSecretName,
-	}
-
-	utils.AddOwnerRefToObject(desired, utils.AsOwner(clusterRequest.Cluster))
-	err := clusterRequest.Create(desired)
-	if err != nil {
-		if !errors.IsAlreadyExists(err) {
-			return fmt.Errorf("Failure creating the collector service: %v", err)
-		}
-		current := &v1.Service{}
-		retryErr := retry.RetryOnConflict(retry.DefaultRetry, func() error {
-			if err = clusterRequest.Get(desired.Name, current); err != nil {
-				if errors.IsNotFound(err) {
-					// the object doesn't exist -- it was likely culled
-					// recreate it on the next time through if necessary
-					return nil
-				}
-				return fmt.Errorf("Failed to get %q service for %q: %v", current.Name, clusterRequest.Cluster.Name, err)
-			}
-			if services.AreSame(current, desired) {
-				log.V(3).Info("Services are the same skipping update")
-				return nil
-			}
-			//Explicitly copying because services are immutable
-			current.Labels = desired.Labels
-			current.Spec.Selector = desired.Spec.Selector
-			current.Spec.Ports = desired.Spec.Ports
-			return clusterRequest.Update(current)
-		})
-		log.V(3).Error(retryErr, "Reconcile Service retry error")
-		return retryErr
-	}
-	return err
-}
-
-func (clusterRequest *ClusterLoggingRequest) reconcileFluentdServiceMonitor() error {
-
-	cluster := clusterRequest.Cluster
-
-	desired := NewServiceMonitor(constants.CollectorName, cluster.Namespace)
-
-	endpoint := monitoringv1.Endpoint{
-		Port:   metricsPortName,
-		Path:   "/metrics",
-		Scheme: "https",
-		TLSConfig: &monitoringv1.TLSConfig{
-			CAFile:     prometheusCAFile,
-			ServerName: fmt.Sprintf("%s.%s.svc", constants.CollectorName, cluster.Namespace),
-		},
-	}
-	logMetricExporterEndpoint := monitoringv1.Endpoint{
-		Port:   exporterPortName,
-		Path:   "/metrics",
-		Scheme: "https",
-		TLSConfig: &monitoringv1.TLSConfig{
-			CAFile:     prometheusCAFile,
-			ServerName: fmt.Sprintf("%s.%s.svc", constants.CollectorName, cluster.Namespace),
-		},
-	}
-
-	labelSelector := metav1.LabelSelector{
-		MatchLabels: map[string]string{
-			"logging-infra": "support",
-		},
-	}
-
-	desired.Spec = monitoringv1.ServiceMonitorSpec{
-		JobLabel:  "monitor-collector",
-		Endpoints: []monitoringv1.Endpoint{endpoint, logMetricExporterEndpoint},
-		Selector:  labelSelector,
-		NamespaceSelector: monitoringv1.NamespaceSelector{
-			MatchNames: []string{cluster.Namespace},
-		},
-	}
-
-	utils.AddOwnerRefToObject(desired, utils.AsOwner(cluster))
-
-	err := clusterRequest.Create(desired)
-	if err != nil {
-		if !errors.IsAlreadyExists(err) {
-			return fmt.Errorf("Failure creating the collector ServiceMonitor: %v", err)
-		}
-		current := &monitoringv1.ServiceMonitor{}
-		retryErr := retry.RetryOnConflict(retry.DefaultRetry, func() error {
-			if err = clusterRequest.Get(desired.Name, current); err != nil {
-				if errors.IsNotFound(err) {
-					// the object doesn't exist -- it was likely culled
-					// recreate it on the next time through if necessary
-					return nil
-				}
-				return fmt.Errorf("Failed to get %q service for %q: %v", current.Name, clusterRequest.Cluster.Name, err)
-			}
-			if servicemonitor.AreSame(current, desired) {
-				log.V(3).Info("ServiceMonitor are the same skipping update")
-				return nil
-			}
-			current.Labels = desired.Labels
-			current.Spec = desired.Spec
-			current.Annotations = desired.Annotations
-
-			return clusterRequest.Update(current)
-		})
-		log.V(3).Error(retryErr, "Reconcile ServiceMonitor retry error")
-		return retryErr
-	}
-	return err
-}
-
-func (clusterRequest *ClusterLoggingRequest) createOrUpdateFluentdPrometheusRule() error {
-	ctx := context.TODO()
-	cluster := clusterRequest.Cluster
-
-	rule := NewPrometheusRule(constants.CollectorName, cluster.Namespace)
-
-	spec, err := NewPrometheusRuleSpecFrom(path.Join(utils.GetShareDir(), fluentdAlertsFile))
-	if err != nil {
-		return fmt.Errorf("failure creating the collector PrometheusRule: %w", err)
-	}
-
-	rule.Spec = *spec
-
-	utils.AddOwnerRefToObject(rule, utils.AsOwner(cluster))
-
-	err = clusterRequest.Create(rule)
-	if err == nil {
-		return nil
-	}
-	if !errors.IsAlreadyExists(err) {
-		return fmt.Errorf("failure creating the collector PrometheusRule: %w", err)
-	}
-
-	current := &monitoringv1.PrometheusRule{}
-	return retry.RetryOnConflict(retry.DefaultRetry, func() error {
-		err = clusterRequest.Client.Get(ctx, types.NamespacedName{Name: rule.Name, Namespace: rule.Namespace}, current)
-		if err != nil {
-			log.V(2).Info("could not get prometheus rule", rule.Name, err)
-			return err
-		}
-		current.Spec = rule.Spec
-		if err = clusterRequest.Client.Update(ctx, current); err != nil {
-			return err
-		}
-		log.V(3).Info("updated prometheus rules")
-		return nil
-	})
-}
-
 // useOldRemoteSyslogPlugin checks if old plugin (docebo/fluent-plugin-remote-syslog) is to be used for sending syslog or new plugin (dlackty/fluent-plugin-remote_syslog) is to be used
 func (clusterRequest *ClusterLoggingRequest) useOldRemoteSyslogPlugin() bool {
 	if clusterRequest.ForwarderRequest == nil {
@@ -290,66 +71,6 @@ func (clusterRequest *ClusterLoggingRequest) useOldRemoteSyslogPlugin() bool {
 	}
 	enabled, found := clusterRequest.ForwarderRequest.Annotations[UseOldRemoteSyslogPlugin]
 	return found && enabled == "enabled"
-}
-
-func (clusterRequest *ClusterLoggingRequest) createOrUpdateFluentdConfigMap(fluentConf string) error {
-	fluentdConfigMap := NewConfigMap(
-		constants.CollectorName,
-		clusterRequest.Cluster.Namespace,
-		map[string]string{
-			"fluent.conf": fluentConf,
-			"run.sh":      fluentd.RunScript,
-		},
-	)
-
-	utils.AddOwnerRefToObject(fluentdConfigMap, utils.AsOwner(clusterRequest.Cluster))
-
-	err := clusterRequest.Create(fluentdConfigMap)
-	if err != nil && !errors.IsAlreadyExists(err) {
-		return fmt.Errorf("Failure constructing collector configmap: %v", err)
-	}
-	retryErr := retry.RetryOnConflict(retry.DefaultRetry, func() error {
-		current := &v1.ConfigMap{}
-		if err = clusterRequest.Get(fluentdConfigMap.Name, current); err != nil {
-			if errors.IsNotFound(err) {
-				log.V(2).Info("Returning nil. The configmap was not found even though create previously failed.  Was it culled?", "configmap name", fluentdConfigMap.Name)
-				return nil
-			}
-			return fmt.Errorf("Failed to get %v configmap for %q: %v", fluentdConfigMap.Name, clusterRequest.Cluster.Name, err)
-		}
-		if reflect.DeepEqual(fluentdConfigMap.Data, current.Data) {
-			return nil
-		}
-		current.Data = fluentdConfigMap.Data
-		return clusterRequest.Update(current)
-	})
-
-	return retryErr
-}
-
-func (clusterRequest *ClusterLoggingRequest) createOrUpdateFluentdSecret() error {
-	var secrets = map[string][]byte{}
-	_ = Syncronize(func() error {
-		secrets = map[string][]byte{}
-		// Ignore errors, these files are optional depending on security settings.
-		secrets["ca-bundle.crt"], _ = ioutil.ReadFile(utils.GetWorkingDirFilePath("ca.crt"))
-		secrets["tls.key"], _ = ioutil.ReadFile(utils.GetWorkingDirFilePath("system.logging.fluentd.key"))
-		secrets["tls.crt"], _ = ioutil.ReadFile(utils.GetWorkingDirFilePath("system.logging.fluentd.crt"))
-		return nil
-	})
-	fluentdSecret := NewSecret(
-		constants.CollectorName,
-		clusterRequest.Cluster.Namespace,
-		secrets)
-
-	utils.AddOwnerRefToObject(fluentdSecret, utils.AsOwner(clusterRequest.Cluster))
-
-	err := clusterRequest.CreateOrUpdateSecret(fluentdSecret)
-	if err != nil {
-		return err
-	}
-
-	return nil
 }
 
 func newFluentdPodSpec(cluster *logging.ClusterLogging, trustedCABundleCM *v1.ConfigMap, pipelineSpec logging.ClusterLogForwarderSpec) v1.PodSpec {
@@ -560,16 +281,9 @@ func newFluentdPodSpec(cluster *logging.ClusterLogging, trustedCABundleCM *v1.Co
 	return fluentdPodSpec
 }
 
-func (clusterRequest *ClusterLoggingRequest) createOrUpdateFluentdDaemonset(pipelineConfHash string) (err error) {
+func (clusterRequest *ClusterLoggingRequest) createOrUpdateFluentdDaemonset(fluentdTrustBundle *v1.ConfigMap, pipelineConfHash string) (err error) {
 
 	cluster := clusterRequest.Cluster
-
-	var fluentdTrustBundle *v1.ConfigMap
-	// Create or update cluster proxy trusted CA bundle.
-	fluentdTrustBundle, err = clusterRequest.createOrGetTrustedCABundleConfigMap(constants.CollectorTrustedCAName)
-	if err != nil {
-		return
-	}
 
 	fluentdPodSpec := newFluentdPodSpec(cluster, fluentdTrustBundle, clusterRequest.ForwarderSpec)
 
@@ -682,7 +396,7 @@ func (clusterRequest *ClusterLoggingRequest) getTrustedCABundleHash() (string, e
 	return trustedCAHashValue, nil
 }
 
-func (clusterRequest *ClusterLoggingRequest) RestartFluentd() (err error) {
+func (clusterRequest *ClusterLoggingRequest) RestartCollector() (err error) {
 
 	collectorConfig, err := clusterRequest.generateCollectorConfig()
 	if err != nil {
@@ -695,12 +409,13 @@ func (clusterRequest *ClusterLoggingRequest) RestartFluentd() (err error) {
 		log.Error(err, "unable to calculate MD5 hash.")
 		return
 	}
+	var collectorType logging.LogCollectionType = clusterRequest.Cluster.Spec.Collection.Logs.Type
 
-	if err = clusterRequest.createOrUpdateFluentdDaemonset(collectorConfHash); err != nil {
+	if err = clusterRequest.createOrUpdateCollectorDaemonset(collectorType, collectorConfHash); err != nil {
 		return
 	}
 
-	return clusterRequest.UpdateFluentdStatus()
+	return clusterRequest.UpdateCollectorStatus(collectorType)
 }
 
 //updateEnvar adds the value to the list or replaces it if it already existing
